+++ mbbsd/talk.c
@@ -2911,7 +2911,7 @@ talkreply(void)
 
     outs("\n\n");
     // FIXME CRASH here
-    assert(sig>=0 && sig<sizeof(sig_des)/sizeof(sig_des[0]));
+    assert(sig>=0 && sig<(int)(sizeof(sig_des)/sizeof(sig_des[0])));
     prints("       (Y) 讓我們 %s 吧！"
 	   "       (A) 我現在很忙，請等一會兒再 call 我\n", sig_des[sig]);
     prints("       (N) 我現在不想 %s "
